Organized distro-bin code into more source files
